# 前端/myoj/openapitools.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "通用后端模版 API 文档",
    "description": "登录、注册、用户管理等接口",
    "contact": {
      "name": "开发团队",
      "email": "dev@example.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8888/api",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/question/admin/update": {
      "post": {
        "tags": [
          "题目接口"
        ],
        "summary": "管理员更新题目",
        "description": "按 id 更新题目",
        "operationId": "adminUpdateQuestion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
    "/question/admin/page": {
      "post": {
        "tags": [
          "题目接口"
        ],
        "summary": "管理员分页查询题目",
        "description": "支持按标题、内容、标签、出题人筛选",
        "operationId": "adminPageQuestions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageQuestionAdminListItemVO"
                }
              }
            }
          }
        }
      }
    },
    "/question/admin/delete/batch": {
      "post": {
        "tags": [
          "题目接口"
        ],
        "summary": "管理员批量删除题目",
        "description": "支持单个与批量删除",
        "operationId": "adminBatchDeleteQuestions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionBatchDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/contest/join": {
      "post": {
        "tags": [
          "contest-controller"
        ],
        "operationId": "join",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContestJoinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/contest/admin/update": {
      "post": {
        "tags": [
          "contest-admin-controller"
        ],
        "operationId": "update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContestAdminUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/contest/admin/page": {
      "post": {
        "tags": [
          "contest-admin-controller"
        ],
        "operationId": "page",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContestAdminPageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageContestAdminListItemVO"
                }
              }
            }
          }
        }
      }
    },
    "/contest/admin/delete/batch": {
      "post": {
        "tags": [
          "contest-admin-controller"
        ],
        "operationId": "deleteBatch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContestAdminBatchDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/contest/admin/add": {
      "post": {
        "tags": [
          "contest-admin-controller"
        ],
        "operationId": "add",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContestAdminAddRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLong"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/admin/page": {
      "get": {
        "tags": [
          "用户管理"
        ],
        "summary": "管理员分页用户列表",
        "description": "仅管理员可访问",
        "operationId": "adminPageUsers",
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 10
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultUserAdminPageVO"
                }
              }
            }
          }
        }
      }
    },
...
    "/question_submit/admin/page": {
      "get": {
        "tags": [
          "question-submit-controller"
        ],
        "operationId": "pageAllQuestionSubmitsForAdmin",
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 10
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "questionId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "submitNo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageMySubmitItemVO"
                }
              }
            }
          }
        }
      }
    },
...
    "/contest/rank": {
      "get": {
        "tags": [
          "contest-controller"
        ],
        "operationId": "rank",
        "parameters": [
          {
            "name": "contestId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListContestRankRowVO"
                }
              }
            }
          }
        }
      }
    },
...
    "/contest/get": {
      "get": {
        "tags": [
          "contest-controller"
        ],
        "operationId": "getContest",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultContestDetailVO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "QuestionQueryRequest": {
        "type": "object",
        "properties": {
          "current": {
            "type": "integer",
            "format": "int64"
          },
          "pageSize": {
            "type": "integer",
            "format": "int64"
          },
          "sortField": {
            "type": "string"
          },
          "sortOrder": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "content": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "userNickname": {
            "type": "string"
          }
        }
      },
      "JudgeCase": {
        "type": "object",
        "properties": {
          "input": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 0
          },
          "output": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 0
          }
        },
        "required": [
          "input",
          "output"
        ]
      },
      "JudgeConfig": {
        "type": "object",
        "properties": {
          "timeLimit": {
            "type": "integer",
            "format": "int64",
            "maximum": 10000
          },
          "memoryLimit": {
            "type": "integer",
            "format": "int64",
            "maximum": 1048576
          },
          "stackLimit": {
            "type": "integer",
            "format": "int64",
            "maximum": 262144
          }
        },
        "required": [
          "memoryLimit",
          "stackLimit",
          "timeLimit"
        ]
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "column": {
            "type": "string"
          },
          "asc": {
            "type": "boolean"
          }
        }
      },
      "PageQuestionAdminListItemVO": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionAdminListItemVO"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "current": {
            "type": "integer",
            "format": "int64"
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            },
            "writeOnly": true
          },
          "optimizeCountSql": {
            "writeOnly": true
          },
          "searchCount": {
            "writeOnly": true
          },
          "optimizeJoinOfCountSql": {
            "type": "boolean",
            "writeOnly": true
          },
          "maxLimit": {
            "type": "integer",
            "format": "int64",
            "writeOnly": true
          },
          "countId": {
            "type": "string",
            "writeOnly": true
          },
          "pages": {
            "type": "integer",
            "format": "int64",
            "deprecated": true
          }
        }
      },
      "QuestionAdminListItemVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "answer": {
            "type": "string"
          },
          "submitNum": {
            "type": "integer",
            "format": "int32"
          },
          "acceptedNum": {
            "type": "integer",
            "format": "int32"
          },
          "judgeCase": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JudgeCase"
            }
          },
          "judgeConfig": {
            "$ref": "#/components/schemas/JudgeConfig"
          },
          "thumbNum": {
            "type": "integer",
            "format": "int32"
          },
          "favourNum": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "userNickname": {
            "type": "string"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ResultPageQuestionAdminListItemVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageQuestionAdminListItemVO"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "QuestionUpdateRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "content": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 0
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "minItems": 1
          },
          "answer": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 0
          },
          "judgeCase": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JudgeCase"
            },
            "minItems": 1
          },
          "judgeConfig": {
            "$ref": "#/components/schemas/JudgeConfig"
          }
        },
        "required": [
          "answer",
          "content",
          "id",
          "judgeCase",
          "judgeConfig",
          "tags",
          "title"
        ]
      },
      "ResultString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "string"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "QuestionBatchDeleteRequest": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "minItems": 1
          }
        },
        "required": [
          "ids"
        ]
      },
      "ResultBoolean": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ResultLong": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "integer",
            "format": "int64"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ContestJoinRequest": {
        "type": "object",
        "properties": {
          "contestId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ContestAdminUpdateRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContestQuestionUpsertItem"
            }
          }
        },
        "required": [
          "endTime",
          "id",
          "startTime",
          "title"
        ]
      },
      "ContestQuestionUpsertItem": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "string",
            "minLength": 1
          },
          "fullScore": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "questionId"
        ]
      },
      "ContestAdminPageRequest": {
        "type": "object",
        "properties": {
          "current": {
            "type": "integer",
            "format": "int64"
          },
          "pageSize": {
            "type": "integer",
            "format": "int64"
          },
          "sortField": {
            "type": "string"
          },
          "sortOrder": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "ContestAdminListItemVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "phase": {
            "type": "string"
          },
          "questionCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PageContestAdminListItemVO": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContestAdminListItemVO"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "current": {
            "type": "integer",
            "format": "int64"
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            },
            "writeOnly": true
          },
          "optimizeCountSql": {
            "writeOnly": true
          },
          "searchCount": {
            "writeOnly": true
          },
          "optimizeJoinOfCountSql": {
            "type": "boolean",
            "writeOnly": true
          },
          "maxLimit": {
            "type": "integer",
            "format": "int64",
            "writeOnly": true
          },
          "countId": {
            "type": "string",
            "writeOnly": true
          },
          "pages": {
            "type": "integer",
            "format": "int64",
            "deprecated": true
          }
        }
      },
      "ResultPageContestAdminListItemVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageContestAdminListItemVO"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ContestAdminBatchDeleteRequest": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "minItems": 1
          }
        },
        "required": [
          "ids"
        ]
      },
      "ContestAdminAddRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContestQuestionUpsertItem"
            }
          }
        },
        "required": [
          "endTime",
          "startTime",
          "title"
        ]
      },
...
      "ResultUserAdminPageVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/UserAdminPageVO"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "UserAdminPageVO": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserAdminListItemVO"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "current": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "PageMySubmitItemVO": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MySubmitItemVO"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "current": {
            "type": "integer",
            "format": "int64"
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            },
            "writeOnly": true
          },
          "optimizeCountSql": {
            "writeOnly": true
          },
          "searchCount": {
            "writeOnly": true
          },
          "optimizeJoinOfCountSql": {
            "type": "boolean",
            "writeOnly": true
          },
          "maxLimit": {
            "type": "integer",
            "format": "int64",
            "writeOnly": true
          },
          "countId": {
            "type": "string",
            "writeOnly": true
          },
          "pages": {
            "type": "integer",
            "format": "int64",
            "deprecated": true
          }
        }
      },
      "ResultPageMySubmitItemVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageMySubmitItemVO"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ContestRankRowVO": {
        "type": "object",
        "properties": {
          "rankOrder": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "userName": {
            "type": "string"
          },
          "solvedCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalPenalty": {
            "type": "integer",
            "format": "int32"
          },
          "totalScore": {
            "type": "integer",
            "format": "int32"
          },
          "totalTime": {
            "type": "integer",
            "format": "int64"
          },
          "totalMemory": {
            "type": "integer",
            "format": "int64"
          },
          "cells": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContestRankCellVO"
            }
          }
        }
      },
      "ResultListContestRankRowVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContestRankRowVO"
            }
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ContestDetailVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "phase": {
            "type": "string"
          },
          "meRegistered": {
            "type": "boolean"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContestQuestionBriefVO"
            }
          }
        }
      },
      "ContestQuestionBriefVO": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "integer",
            "format": "int64"
          },
          "questionTitle": {
            "type": "string"
          },
          "fullScore": {
            "type": "integer",
            "format": "int32"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ResultContestDetailVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ContestDetailVO"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  }
}